requirements.txt
Django==6.0.3
Pillow==12.1.1
djangorestframework==3.17.1
django-filter==25.2

# Testing
pytest==9.0.2
pytest-django==4.12.0
